feat: refactor for eventual SVM support
